pyproject.toml
[project]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "greenlet (>=3.3.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
